+++ app/src/main/java/com/example/valorant_app/ui/ValorantWikiApp.kt
@@ -1,21 +1,9 @@
 package com.example.valorant_app.ui
 
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.stringResource
-import androidx.navigation.NavController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.valorant_app.R
 import com.example.valorant_app.ui.navigation.AgentRoute
 import com.example.valorant_app.ui.navigation.HomePageRoute
 import com.example.valorant_app.ui.navigation.InitialPageRoute
@@ -26,8 +14,12 @@ import com.example.valorant_app.ui.pages.home.HomeContent
 import com.example.valorant_app.ui.pages.initial_screen.InitialScreen
 import com.example.valorant_app.ui.pages.weapon.card.WeaponSkinsScreen
 import com.example.valorant_app.ui.pages.weapon.single.WeaponSingleScreen
+import com.example.valorant_app.ui.reusable_comp.AgentSingleTopBar
+import com.example.valorant_app.ui.reusable_comp.AgentTopBar
 import com.example.valorant_app.ui.reusable_comp.AppScaffold
-import com.example.valorant_app.ui.reusable_comp.FlagsDropdown
+import com.example.valorant_app.ui.reusable_comp.HomeTopBar
+import com.example.valorant_app.ui.reusable_comp.WeaponSingleTopBar
+import com.example.valorant_app.ui.reusable_comp.WeaponTopBar
 
 @Composable
 fun ValorantWikiApp() {
@@ -94,75 +86,3 @@ fun ValorantWikiApp() {
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun HomeTopBar(navController: NavController) {
-    TopAppBar(
-        title = { Text(stringResource(R.string.app_name), color = Color.White) },
-        colors = TopAppBarDefaults.topAppBarColors(
-            containerColor = Color(0xFF0E0E10),
-            titleContentColor = Color(0xFFE03240)
-        ),
-        actions = { FlagsDropdown() }
-    )
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun AgentTopBar(navController: NavController) {
-    TopAppBar(
-        title = { Text(stringResource(R.string.agents), color = Color.White) },
-        colors = TopAppBarDefaults.topAppBarColors(
-            containerColor = Color(0xFF0E0E10),
-            titleContentColor = Color(0xFFE03240)
-        )
-    )
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun WeaponTopBar(navController: NavController) {
-    TopAppBar(
-        title = { Text(stringResource(R.string.weapon), color = Color.White) },
-        colors = TopAppBarDefaults.topAppBarColors(
-            containerColor = Color(0xFF0E0E10),
-            titleContentColor = Color(0xFFE03240)
-        )
-    )
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun AgentSingleTopBar(navController: NavController) {
-    TopAppBar(
-        title = { Text(stringResource(R.string.agent_details), color = Color.White) },
-        navigationIcon = {
-            IconButton(onClick = { navController.popBackStack() }) {
-                Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = null, tint = Color.White)
-            }
-        },
-        colors = TopAppBarDefaults.topAppBarColors(
-            containerColor = Color(0xFF0E0E10),
-            titleContentColor = Color(0xFFE03240),
-            navigationIconContentColor = Color.White
-        )
-    )
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun WeaponSingleTopBar(navController: NavController) {
-    TopAppBar(
-        title = { Text(stringResource(R.string.weapon_skins), color = Color.White) },
-        navigationIcon = {
-            IconButton(onClick = { navController.popBackStack() }) {
-                Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = null, tint = Color.White)
-            }
-        },
-        colors = TopAppBarDefaults.topAppBarColors(
-            containerColor = Color(0xFF0E0E10),
-            titleContentColor = Color(0xFFE03240),
-            navigationIconContentColor = Color.White
-        )
-    )
-}
