+++ app/src/main/java/com/example/valorant_app/ui/reusable_comp/TopAppBarViewModel.kt
@@ -1,0 +1,8 @@
+package com.example.valorant_app.ui.reusable_comp
+
+import androidx.lifecycle.ViewModel
+
+class TopAppBarViewModel : ViewModel() {
+
+
+}
